// src/app/shared/models/flightcriteria.model.ts
export class Flightcriteria {
    company: string;
    flightType: string;
    travelType: string;
    departureLocation: string;
    arrivalLocation: string;
    departureDateMin: Date;
    arrivalDateMin: Date;
    backDateMin: Date;
    departureTimeMin: Date;
    arrivalTimeMin: Date;
    backTimeMin: Date;
    flightDurationMin: Date;
    connectionDurationMin: Date;
    departureDateMax: Date;
    arrivalDateMax: Date;
    backDateMax: Date;
    departureTimeMax: Date;
    arrivalTimeMax: Date;
    backTimeMax: Date;
    flightDurationMax: Date;
    connectionDurationMax: Date;
    aircraftType: string;
    fareMin: number;
    fareMax: number;
}